Chatbot Model SRV/requirements.txt
beautifulsoup4==4.12.2
einops==0.6.1
Flask==2.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
huggingface-hub==0.15.1
importlib-metadata==6.6.0
jedi 
loralib==0.1.1
matplotlib-inline 
multiprocess==0.70.14
nest-asyncio 
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
pandas==2.0.2
peft 
regex==2023.6.3
requests==2.31.0
responses==0.18.0
scipy==1.10.1
sympy==1.12
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
transformers 
urllib3==2.0.2
asyncio-nats-streaming
pymongo
